Optimize id = id and id != id nodes
When id (placeholder name) has a classic non-nullable type, we can
replace the node with true or false. When id has a classic nullable
type, we can replace the node with isNull(id) or isNotNull(id). For
other types that cannot be nullable (Nothing, for example), we don't
optimize the node.